subdomain-takeover-scanner/requirements.txt
dnspython>=2.6.1
httpx>=0.27.0
